requirements.txt
Django
django-rest-framework
django-cors-headers
psycopg2
python-dotenv
django-phonenumber-field
Pillow
phonenumbers
djangorestframework-simplejwt
geopy
requests